Scaffold plugin tests.
This commit uses the scaffolding provided by WP-CLI, with a few modifications to the Travis CI configuration (only run on PHP 7.0+, and ignore PHP_CodeSniffer for now). It also introduces some tests around the creation of new users, since that's kind of an important piece of functionality for the plugin.